Relocated source code to their own header file
